dat phong

+++ backend/src/main/java/com/example/demo/Entities/HoaDon.java
@@ -15,7 +15,7 @@ import lombok.AllArgsConstructor;
 public class HoaDon {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long hoaDonID;
+    private Integer hoaDonID;
 
     @ManyToOne
     @JoinColumn(name = "maNguoiDung",referencedColumnName = "maNguoiDung")
